# main-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Forest Run API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/runs/Invite": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Invite run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Invite run",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.InviteRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.InviteRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/runs/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Create run",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.CreateRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.CreateRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/runs/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Update run",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.UpdateRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.UpdateRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/subscribe": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Start receiving notifications of user's actions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Subscribe to user",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.SubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/unsubscribe": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Stop receiving notifications of user's actions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Unsubscribe from user",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.SubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Refresh JWT Tokens pair",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Tokens",
                "parameters": [
                    {
                        "description": "Refresh tokens data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.RefreshTokensRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.RefreshTokensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "Log into app and get access tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign In",
                "parameters": [
                    {
                        "description": "SignIn data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Register in app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign Up",
                "parameters": [
                    {
                        "description": "SignUp data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.SignUpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.JWTTokenData": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "domain.Run": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "creator": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_participants": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "official_site_url": {
                    "type": "string"
                },
                "participation_format": {
                    "$ref": "#/definitions/domain.RunParticipationFormat"
                },
                "permissions": {
                    "$ref": "#/definitions/runs.PermissionsType"
                },
                "registration_until": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/runs.Route"
                },
                "start_place": {
                    "$ref": "#/definitions/runs.Place"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.RunStatus"
                }
            }
        },
        "domain.RunParticipationFormat": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "OpenRunFormat",
                "ClosedRunFormat"
            ]
        },
        "domain.RunStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                4
            ],
            "x-enum-varnames": [
                "OpenRunStatus",
                "RunningRunStatus",
                "FinishedRunStatus",
                "CanceledRunStatus"
            ]
        },
        "protocol.CreateRunRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "max_participants": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "official_site_url": {
                    "type": "string"
                },
                "participation_format": {
                    "$ref": "#/definitions/domain.RunParticipationFormat"
                },
                "permissions": {
                    "$ref": "#/definitions/runs.PermissionsType"
                },
                "registration_until": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/runs.Route"
                },
                "start_place": {
                    "$ref": "#/definitions/runs.Place"
                },
                "start_time": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "protocol.CreateRunResponse": {
            "type": "object",
            "properties": {
                "run": {
                    "$ref": "#/definitions/domain.Run"
                }
            }
        },
        "protocol.ErrorCode": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "CodeUserNameAlreadyTaken",
                "CodeWrongPassword"
            ]
        },
        "protocol.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/protocol.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "protocol.InviteRunRequest": {
            "type": "object",
            "properties": {
                "run_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "protocol.InviteRunResponse": {
            "type": "object"
        },
        "protocol.RefreshTokensRequest": {
            "type": "object",
            "properties": {
                "authToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "protocol.RefreshTokensResponse": {
            "type": "object",
            "properties": {
                "authToken": {
                    "$ref": "#/definitions/domain.JWTTokenData"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "protocol.SignInRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "protocol.SignInResponse": {
            "type": "object",
            "properties": {
                "authToken": {
                    "$ref": "#/definitions/domain.JWTTokenData"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "protocol.SignUpRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "protocol.SignUpResponse": {
            "type": "object"
        },
        "protocol.SubscriptionRequest": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "integer"
                }
            }
        },
        "protocol.SubscriptionResponse": {
            "type": "object"
        },
        "protocol.UpdateRunRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "max_participants": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "official_site_url": {
                    "type": "string"
                },
                "participation_format": {
                    "$ref": "#/definitions/domain.RunParticipationFormat"
                },
                "permissions": {
                    "$ref": "#/definitions/runs.PermissionsType"
                },
                "registration_until": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/runs.Route"
                },
                "run_id": {
                    "type": "integer"
                },
                "start_place": {
                    "$ref": "#/definitions/runs.Place"
                },
                "start_time": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "protocol.UpdateRunResponse": {
            "type": "object",
            "properties": {
                "run": {
                    "$ref": "#/definitions/domain.Run"
                }
            }
        },
        "runs.PermissionsType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "FreePermissionsType"
            ]
        },
        "runs.Place": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "point": {
                    "$ref": "#/definitions/runs.Point"
                }
            }
        },
        "runs.Point": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "runs.Route": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/runs.Point"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}